1.Navbar hilighting moved from server to client. 2.Add admin page update config authentication.

+++ update/updateaddresses.php
@@ -1,16 +1,19 @@
-<?
+<?php
 /**
  * @author Greedi
  * @copyright 2012
  */
 include('../core/wallet.php');
 
-$dailytotal = $_POST['delete'];
-
+if(is_admin()){
+    $dailytotal = $_POST['delete'];
     mysql_query("DELETE FROM dailyltc") 
     or die(mysql_error());
-    
-    header( 'Location: http://d.evco.in/faucet/server.php' ) ;  
+    header( 'Location: /server.php' );
+}else{
+    echo "Access Denied.";
+}
+
 ?>
